requirements.txt
cassandra-driver==2.6.0
requests==2.7.0